move alfred to the new filesystem format
