--- alfred-names/soul-engine.json
@@ -1,4 +1,0 @@
-{
-  "path": "src",
-  "soul": "alfred-learns"
-}
