+++ alfred-names/soul/mentalProcesses/brainstorms.ts
