+++ alfred-names/soul/subProcesses/compress.ts
